Move BorrowerModel to file
