+++ lib/src/features/borrowers/data/borrowers_view_model.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:librarian_app/src/features/common/data/lending_api.dart';
 
+import 'borrower_model.dart';
 import 'borrowers_mapper.dart';
 
 class BorrowersViewModel extends ChangeNotifier {
@@ -69,45 +70,4 @@ class BorrowersViewModel extends ChangeNotifier {
     refresh();
     return true;
   }
-}
-
-class BorrowerModel {
-  final String id;
-  final String name;
-  final String? email;
-  final String? phone;
-  final List<Issue> issues;
-
-  bool get active => issues.isEmpty;
-
-  BorrowerModel({
-    required this.id,
-    required this.name,
-    required this.issues,
-    this.email,
-    this.phone,
-  });
-}
-
-class Issue {
-  final IssueType type;
-  final String title;
-  final String? explanation;
-  final String? instructions;
-  final String? graphicUrl;
-
-  const Issue({
-    required this.title,
-    this.explanation,
-    this.instructions,
-    this.graphicUrl,
-    required this.type,
-  });
-}
-
-enum IssueType {
-  duesNotPaid,
-  overdueLoan,
-  suspended,
-  needsLiabilityWaiver,
 }
