+++ lib/src/features/borrowers/presentation/borrower_details.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:librarian_app/src/features/borrowers/presentation/borrower_issues.dart';
 
-import '../data/borrowers_view_model.dart';
+import '../data/borrower_model.dart';
 
 class BorrowerDetails extends StatefulWidget {
   final BorrowerModel borrower;
