+++ lib/src/features/borrowers/presentation/borrower_details_pane.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:librarian_app/src/features/borrowers/presentation/borrower_details.dart';
 
-import '../data/borrowers_view_model.dart';
+import '../data/borrower_model.dart';
 
 class BorrowerDetailsPane extends StatelessWidget {
   final BorrowerModel? borrower;
