+++ lib/src/features/borrowers/presentation/searchable_borrowers_list.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:librarian_app/src/features/borrowers/data/borrowers_view_model.dart';
+import 'package:librarian_app/src/features/borrowers/data/borrower_model.dart';
 import 'package:librarian_app/src/features/borrowers/presentation/connected_borrowers_list.dart';
 
 import '../../common/presentation/submit_text_field.dart';
