+++ lib/src/features/loans/data/wizard_model.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:librarian_app/src/features/borrowers/data/borrowers_view_model.dart';
+import 'package:librarian_app/src/features/borrowers/data/borrower_model.dart';
 import 'package:librarian_app/src/features/loans/data/things_model.dart';
 
 class WizardModel extends ChangeNotifier {
